requirements.txt
fastapi~=0.109.0
uvicorn~=0.27.0
mysql-connector-python~=8.0.33
pydantic~=2.5.3
black
flake8
pandas~=2.2.0
openpyxl
Pyarrow
numpy~=1.26.3
python-multipart
starlette~=0.35.1
SQLAlchemy~=2.0.25
pylint
